backend/requirements.txt
Flask==3.0.3
Flask-CORS==4.0.1
Flask-JWT-Extended==4.6.0
bcrypt==4.1.2
Flask-SQLAlchemy==3.1.1
flask-marshmallow==1.3.0
marshmallow-sqlalchemy==1.0.0
python-dotenv==1.0.1
Flask-Migrate==4.0.5
